+++ src/main/java/lotto/view/OutputView.java
@@ -1,12 +1,12 @@
 package lotto.view;
 
-import static lotto.view.OutputMessages.OUTPUT_FIFTH_RESULT_FORMAT;
-import static lotto.view.OutputMessages.OUTPUT_FIRST_RESULT_FORMAT;
-import static lotto.view.OutputMessages.OUTPUT_FORTH_RESULT_FORMAT;
-import static lotto.view.OutputMessages.OUTPUT_PROFIT_RESULT_FORMAT;
-import static lotto.view.OutputMessages.OUTPUT_SECOND_RESULT_FORMAT;
-import static lotto.view.OutputMessages.OUTPUT_THIRD_RESULT_FORMAT;
-import static lotto.view.OutputMessages.OUTPUT_WIN_STATISTICS_MESSAGES;
+import static lotto.view.messages.OutputMessages.OUTPUT_FIFTH_RESULT_FORMAT;
+import static lotto.view.messages.OutputMessages.OUTPUT_FIRST_RESULT_FORMAT;
+import static lotto.view.messages.OutputMessages.OUTPUT_FORTH_RESULT_FORMAT;
+import static lotto.view.messages.OutputMessages.OUTPUT_PROFIT_RESULT_FORMAT;
+import static lotto.view.messages.OutputMessages.OUTPUT_SECOND_RESULT_FORMAT;
+import static lotto.view.messages.OutputMessages.OUTPUT_THIRD_RESULT_FORMAT;
+import static lotto.view.messages.OutputMessages.OUTPUT_WIN_STATISTICS_MESSAGES;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -18,7 +18,7 @@ import lotto.domain.lotto.LottoTicket;
 public class OutputView {
 
     public static void printResult(LottoResult lottoResult) {
-        System.out.println(OUTPUT_WIN_STATISTICS_MESSAGES.getMessage());
+        System.out.println(OUTPUT_WIN_STATISTICS_MESSAGES);
         System.out.printf((OUTPUT_FIFTH_RESULT_FORMAT.getMessage()) + "%n", lottoResult.getNumberOfRank(Rank.FIFTH));
         System.out.printf((OUTPUT_FORTH_RESULT_FORMAT.getMessage()) + "%n", lottoResult.getNumberOfRank(Rank.FOURTH));
         System.out.printf((OUTPUT_THIRD_RESULT_FORMAT.getMessage()) + "%n", lottoResult.getNumberOfRank(Rank.THIRD));
